WILD6900: Course materials for WILD6900: Applied Bayesian Analysis for Ecological Data

Course materials for WILD6900: Applied Bayesian Analysis for Ecological Data at Utah State University. This package primarily exists to distribute data, code, and other materials in an efficient manner.

Author

Maintainer: Clark Rushing clark.rushing@usu.edu