Breeding Bird Survey data

A dataset containing the raw BBS counts and covariate data for BCRs 9, 10, and 16 (1966-present).

bbs_data

Format

A list with 3 data frames:

routes

ROUTE ATTRIBUTES

`countrynum` The three digit code that identifies country where route is located `statenum` The two digit numerical code that identifies the state, province, or territory where the route is located `Route` The three digit code that identifies the route `Active` 1 if the route is currently active, 0 if it is not `Latitude` The latitude of the route start point in decimal degrees `Longitude` The longitude of the route start point in decimal degrees `BCR` The Bird Conservation Region where the route is located `RouteTypeID` Indicates route length and selection criteria (1=random, 50 Stops, 2 =Not Random, 50 Stops, 3=Not Random, < 50 stops) `routeID` Unique eight digit route ID
weather

RUN ATTRIBUTES

`RPID` The run protocol identification number. `Year` The year `Month` The month the route was run (1-12) `Day` The day the route was run (1-31) `ObsN` Unique observer identification number `StartTemp` The temperature recorded at the beginning of the run of the route `EndTemp` The temperature recorded at the end of the run of the route `StartWind` The Beaufort wind speed code recorded at the beginning of the run of the route `EndWind` The Beaufort wind speed code recorded at the end of the run of the route `StartSky` The Weather Bureau sky code recorded at the beginning of the run of the route `EndSky` The Weather Bureau sky code recorded at the end of the run of the route `StartTime` The time the run of the route began, recorded in 24 hour local time `EndTime` The time the run of the route ended, recorded in 24 hour local time `Assistant` If someone assisted the observer by recording the data `RunType` The RunType code helps to quickly determine which data do, or do not meet the BBS program’s data criteria. A RunType code of 1 is assigned whenever data were collected under conditions that meet BBS weather, date, time, and route completion criteria (QualityCurrentID = 1) on a randomly established route (i.e., RouteTypeDetailID = 1) using the official BBS sampling protocol (RunProtocolID = 101) `routeID` Unique eight digit route ID
counts

SPECIES COUNTS

`Year` The year `routeID` Unique eight digit route ID `aou` All remaining columns are the counts of specific species; column names refer to the aou code used by BBS to identify species; values indicate the total number of individuals counted at each route in each year

Source

https://www.mbr-pwrc.usgs.gov/bbs/