sim_Nmix

Simulate data for the N-mixture model with site-level covariats on lambda and occasion-level covariates on p

sim_Nmix(J = 200, K = 3, alpha = c(1, 3), beta = c(0, -5))

Arguments

J:
number of sites at which counts were made

K:
number of times that counts were made at each site

alpha:
vector containing the intercept and slopes of log-linear regression relating abundance to the site covariate A

beta:
ivector containing the intercept and slopes of logistic-linear regression of detection probability on B

Value

List containing the simulated data (y, XN, Xp) and the data-generating values